Dao, Entity 정리
