+++ app/src/main/java/com/example/kotlincalendar/FrdView_rec_Adapter.kt
@@ -7,8 +7,8 @@ import android.view.ViewGroup
 import android.widget.BaseAdapter
 import android.widget.Toast
 import com.example.kotlincalendar.database.AppDatabase
-import com.example.kotlincalendar.database.frdadd_db
-import com.example.kotlincalendar.database.frdlist_db
+import com.example.kotlincalendar.Entity.FriendAdd
+import com.example.kotlincalendar.Entity.FriendList
 import com.example.kotlincalendar.databinding.ListFriendRecItemBinding
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -18,7 +18,7 @@ import kotlinx.coroutines.withContext
 
 class FrdView_rec_Adapter(
     val context: Context,
-    val frdRequests:List<frdadd_db>
+    val frdRequests:List<FriendAdd>
     ) : BaseAdapter() {
     //private var mBinding: ListFriendRecItemBinding? = null
     //private val binding get() = mBinding!!
@@ -56,10 +56,10 @@ class FrdView_rec_Adapter(
             }
         }
         acc_btn.setOnClickListener(){
-            val friendlistAdd= frdlist_db(0, User1=senderEmail, User2=receiverEmail)
+            val friendlistAdd= FriendList(0, User1=senderEmail, User2=receiverEmail)
             CoroutineScope(Dispatchers.IO).launch {
-                db?.frdlistDbDao()?.insertFriend(friendlistAdd)
-                db?.frdaddDbDao()?.deleteRequest(frdRequests[position])
+                db?.friendListDao()?.insertFriend(friendlistAdd)
+                db?.friendAddDao()?.deleteRequest(frdRequests[position])
                 withContext((Dispatchers.Main)){
                     notifyDataSetChanged()
                     Toast.makeText(context, "신청을 수락했습니다", Toast.LENGTH_SHORT).show()
@@ -68,7 +68,7 @@ class FrdView_rec_Adapter(
         }
         dis_btn.setOnClickListener(){
             CoroutineScope(Dispatchers.IO).launch{
-                db?.frdaddDbDao()?.deleteRequest(frdRequests[position])
+                db?.friendAddDao()?.deleteRequest(frdRequests[position])
                 withContext(Dispatchers.Main){
                     notifyDataSetChanged()
                     Toast.makeText(context,"신청을 거절했습니다",Toast.LENGTH_SHORT).show()
