+++ app/src/main/java/com/example/kotlincalendar/database/AppDatabase.kt
@@ -5,14 +5,23 @@ import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.example.kotlincalendar.Dao.FriendAddDao
+import com.example.kotlincalendar.Dao.FriendListDao
+import com.example.kotlincalendar.Dao.UserCalendarDao
+import com.example.kotlincalendar.Dao.UserDao
+import com.example.kotlincalendar.Entity.FriendAdd
+import com.example.kotlincalendar.Entity.FriendList
+import com.example.kotlincalendar.Entity.UserCalendar
+import com.example.kotlincalendar.Entity.User
 
-@Database(entities = [User::class, frdadd_db::class, frdlist_db::class, UserCalendar::class], version = 1)
+
+@Database(entities = [User::class, FriendAdd::class, FriendList::class, UserCalendar::class], version = 1)
 @TypeConverters(LocalDateConverter::class, LocalTimeConverter::class) // TypeConverter 등록
 abstract class AppDatabase: RoomDatabase() {
-    abstract fun userDao(): User_Dao
-    abstract fun frdaddDbDao(): frdadd_db_Dao
-    abstract fun frdlistDbDao(): frdlist_db_Dao
-    abstract fun userCalendarDao(): UserCalendar_Dao
+    abstract fun userDao(): UserDao
+    abstract fun friendAddDao(): FriendAddDao
+    abstract fun friendListDao(): FriendListDao
+    abstract fun userCalendarDao(): UserCalendarDao
 
     companion object {
         private var instance: AppDatabase? = null
